Create an extension class for MethodDeclarationSyntax to share common code
